Normal login works
- created user class
- implemented normal login

Right now the normal login will send back an error message if the login is unsuccessful. These should be interpreted by the frontend to decide how to proceed.
For example, an error message of mustSignup should probably redirect the user to the signup page. An error message of googleuser should alert the user that they signed up through google and should use that to sign in